requirements.txt
openai-agents
gradio
python-dotenv
requests
sendgrid